Added !=, >> operators

+++ 3_uzduotis/Functions.h
@@ -60,7 +60,7 @@ void random_num_generator(vector<double>&, double&);
 
 
 template <typename T> void Rankinis_ivedimas(T&, string&, string&, vector<string>&, vector<string>&);
-template <typename T> T cin_and_checkFormat_in_interval(int, int);
+template <typename T> T cin_and_checkFormat_in_interval(double, double);
 //template <typename T> void Failo_nuskaitymas(string, string, string, int);
 template <typename T> void Failo_skirstymas(int, T&);
 template <typename T> void writeToFile(T&, string);
@@ -111,7 +111,7 @@ bool cool_students_sort(const Student &n) {
 
 
 
-template <typename T> T cin_and_checkFormat_in_interval(int a, int b) {
+template <typename T> T cin_and_checkFormat_in_interval(double a, double b) {
 	T input;
 	while (!(cin >> input) || a > input || input > b) {
 		std::cout << "Netinkamas ivesties formatas. Iveskite is naujo: " << endl;
